+++ R/BatchtoolsFuture-class.R
@@ -1,5 +1,7 @@
 #' A batchtools future is a future whose value will be resolved via batchtools
 #'
+#' @inheritParams future::FutureBackend
+#'
 #' @param resources (optional) A named list passed to the \pkg{batchtools}
 #' template (available as variable `resources`).  See Section 'Resources'
 #' in [batchtools::submitJobs()] more details.
@@ -39,6 +41,7 @@ BatchtoolsFutureBackend <- function(workers = NULL, resources = list(),
                              conf.file = findConfFile(),
                              cluster.functions = NULL,
                              registry = list(),
+                             interrupts = TRUE,
                              ...) {
   assert_no_positional_args_but_first()
 
@@ -79,7 +82,7 @@ BatchtoolsFutureBackend <- function(workers = NULL, resources = list(),
   }
   
   stop_if_not(is.list(resources))
-
+  
   core <- FutureBackend(
     reg = "workers-batchtools",
     workers = workers,
@@ -88,6 +91,7 @@ BatchtoolsFutureBackend <- function(workers = NULL, resources = list(),
     cluster.functions = cluster.functions,
     registry = registry,
     finalize = finalize,
+    interrupts = interrupts,
     future.wait.timeout = getOption("future.wait.timeout", 30 * 24 * 60 * 60),
     future.wait.interval = getOption("future.wait.interval", 0.01),
     future.wait.alpha = getOption("future.wait.alpha", 1.01),
@@ -343,7 +347,12 @@ interruptFuture.BatchtoolsFutureBackend <- function(backend, future, ...) {
       mdebug_pop()
     })
   }
-    
+
+  if (!backend[["interrupts"]]) {
+    if (debug) mdebug("Interrupts are disabled for the current backend")
+    return(future)
+  }
+  
   config <- future[["config"]]
   reg <- config[["reg"]]